Apply hero transition between AgentsScreen and ProfilePage screen
